% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSigma.R
\name{BuildSigma}
\alias{BuildSigma}
\title{Covariance Tensor between Covariance Matrices}
\usage{
BuildSigma(matrix.array)
}
\arguments{
\item{matrix.array}{m x m x k array of k covariance matrices with m traits}
}
\value{
projected covariance tensor: a square symmetric matrix with m(m + 1)/2 rows
}
\description{
Computes the two-dimensional projection of the fourth-order 
covariance tensor between covariance matrices.
}
\author{
Guilherme Garcia
}
\references{
Basser P. J., Pajevic S. 2007. Spectral decomposition of a 4th-order 
covariance tensor: Applications to diffusion tensor MRI. Signal Processing. 87:220–236.

Hine E., Chenoweth S. F., Rundle H. D., Blows M. W. 2009. Characterizing 
the evolution of genetic variance using genetic covariance tensors. Philosophical 
transactions of the Royal Society of London. Series B, Biological sciences. 364:1567–78.
}
\seealso{
\code{\link{EigenTensorDecomposition}}
}

