% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localShapeVariables.R
\name{Center2MeanJacobianFast}
\alias{Center2MeanJacobianFast}
\title{Centered jacobian residuals}
\usage{
Center2MeanJacobianFast(jacobArray)
}
\arguments{
\item{jacobArray}{Arrays of Jacobian calculated in the JacobianArray function}
}
\value{
array of centered residual jacobians
}
\description{
Calculates mean jacobian matrix for a set of jacobian matrices
describing a local aspect of shape deformation for a given set of volumes,
returning log determinants of deviations from mean jacobian (Woods, 2003).
}
\references{
Woods, Roger P. 2003. “Characterizing Volume and Surface
Deformations in an Atlas Framework: Theory, Applications, and Implementation.” NeuroImage 18 (3):769-88.
}
\author{
Guilherme Garcia

Diogo Melo
}
