% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvolQG.R
\docType{data}
\name{dentus}
\alias{dentus}
\title{Example multivariate data set}
\format{
A data frame with 300 rows and 5 variables
}
\usage{
data(dentus)
}
\description{
Simulated example of 4 continuous bone lengths from 5 species.
}
\details{
\itemize{
\item humerus 
\item ulna 
\item femur 
\item tibia 
\item species 
}
}
\keyword{datasets}
