% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvolQG.R
\docType{data}
\name{dentus.tree}
\alias{dentus.tree}
\title{Tree for dentus example species}
\format{
ape tree object
}
\usage{
data(dentus.tree)
}
\description{
Hypothetical tree for the species in the dentus data set.
}
\keyword{datasets}
