% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DriftTest.R
\name{DriftTest}
\alias{DriftTest}
\title{Test drift hypothesis}
\usage{
DriftTest(means, cov.matrix, verbose = TRUE)
}
\arguments{
\item{means}{list or array of species means being compared. array must have means in the rows.}

\item{cov.matrix}{ancestral covariance matrix for all populations}

\item{verbose}{If TRUE all results are returned, is FALSE only regression coeficient}
}
\description{
Given a set of covariance matrices and means for terminals, test the hypothesis
that obseved divergency is larger/smaller than expected by drift alone.
}
\author{
Ana Paula Assis, Diogo Melo
}
\references{
Marroig, G., & Cheverud, J. M. (2004). Did natural selection or genetic drift
produce the cranial diversification of neotropical monkeys? The American Naturalist, 163(3), 417-428. doi:10.1086/381693
}

