% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenTensorDecomposition.R
\name{EigenTensorDecomposition}
\alias{EigenTensorDecomposition}
\title{Eigentensor Decomposition}
\usage{
EigenTensorDecomposition(matrix.array, return.projection = FALSE)
}
\arguments{
\item{matrix.array}{k x k x m array of m covariance matrices with k traits;}

\item{return.projection}{Should we project covariance matrices into estimated eigentensors? Defaults to TRUE.}
}
\value{
List with the following components:

values vector of ordered eigenvalues associated with eigentensors;

matrices array of eigentensor in matrix form;

projection matrix of projected covariance matrices over eigentensors.
}
\description{
This function performs eigentensor decomposition on a set of covariance matrices.
}
\details{
The number of estimated eigentensors is the minimum between k(k+1)/2 and m, 
in a similar manner to the usual principal component analysis.
}
\references{
Basser P. J., Pajevic S. 2007. Spectral decomposition of a 4th-order 
covariance tensor: Applications to diffusion tensor MRI. Signal Processing. 87:220–236.

Hine E., Chenoweth S. F., Rundle H. D., Blows M. W. 2009. Characterizing 
the evolution of genetic variance using genetic covariance tensors. Philosophical 
transactions of the Royal Society of London. Series B, Biological sciences. 364:1567–78.
}

