% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvolQG.R
\docType{package}
\name{evolqg}
\alias{evolqg}
\alias{evolqg-package}
\title{EvolQG}
\description{
The package for evolutionary quantitative genetics.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/lem-usp/evolqg/issues}
}

}
\author{
\strong{Maintainer}: Diogo Melo \email{diogro@gmail.com} (\href{https://orcid.org/0000-0002-7603-0092}{ORCID})

Authors:
\itemize{
  \item Ana Paula Assis (\href{https://orcid.org/0000-0001-8170-6352}{ORCID})
  \item Fabio Andrade Machado (\href{https://orcid.org/0000-0002-0215-9926}{ORCID})
  \item Guilherme Garcia
}

Other contributors:
\itemize{
  \item Edgar Zanella [contributor]
  \item Alex Hubbe (\href{https://orcid.org/0000-0002-3226-0144}{ORCID}) [reviewer]
  \item Gabriel Marroig (\href{https://orcid.org/0000-0003-2542-2034}{ORCID}) [thesis advisor]
}

}
