% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localShapeVariables.R
\name{JacobianArray}
\alias{JacobianArray}
\title{Local Jacobian calculation}
\usage{
JacobianArray(spline, tesselation, ...)
}
\arguments{
\item{spline}{Thin plate spline calculated by the TPS function}

\item{tesselation}{matrix of landmarks.}

\item{...}{Additional arguments to some function}
}
\value{
array of jacobians calculated at the centroids
}
\description{
Calculates jacobians for a given interpolation in a set of points
determined from tesselation (as centroids of each tetrahedron defined, for now...)
}
\note{
Jacobians are calculated on the row centroids of the tesselation matrix.
}
\author{
Guilherme Garcia
}
