% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrzSubspaceBootstrap.R
\name{KrzSubspaceDataFrame}
\alias{KrzSubspaceDataFrame}
\title{Extract confidence intervals from KrzSubspaceBootstrap}
\usage{
KrzSubspaceDataFrame(x, n = ncol(observed), prob = 0.95)
}
\arguments{
\item{x}{output from KrzSubspaceBootstrap function.}

\item{n}{number of eigenvalues to use}

\item{prob}{Posterior probability interval. Default is 95\%.}
}
\value{
Posterior intervals for the eigenvalues of the H matrix in the KrzSubspace comparison.
}
\description{
Returns posterior means and confidence intervals from the object produced by the KrzSubspaceBootstrap() function. Mainly used for ploting using PlotKrzSubspace. 
See example in the KrzSubspaceBootstrap function.
}
\seealso{
\code{\link{KrzSubspaceBootstrap}}, \code{\link{PlotKrzSubspace}}
}
