% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MatrixDistance}
\alias{MatrixDistance}
\alias{MatrixDistance.default}
\alias{MatrixDistance.list}
\title{Matrix Distance}
\usage{
MatrixDistance(cov.x, cov.y, Distance, ...)

\method{MatrixDistance}{default}(cov.x, cov.y, Distance = c("OverlapDist",
  "RiemannDist"), ...)

\method{MatrixDistance}{list}(cov.x, cov.y = NULL,
  Distance = c("OverlapDist", "RiemannDist"), ..., num.cores = 1)
}
\arguments{
\item{cov.x}{Single covariance matrix or list of covariance matrices.
If single matrix is suplied, it is compared to cov.y.
If list is suplied and no cov.y is suplied, all matrices
are compared.
If cov.y is suplied, all matrices in list are compared to it.}

\item{cov.y}{First argument is compared to cov.y.
Optional if cov.x is a list.}

\item{Distance}{Distance function for use in calculation. Currently supports "Riemann" and "Overlap".}

\item{...}{aditional arguments passed to other methods}

\item{num.cores}{If list is passed, number of threads to use in computation. Requires doMC library.}
}
\value{
If cov.x and cov.y are passed, returns distance between them.

If is a list cov.x and cov.y are passed, same as above, but for all matrices in cov.x.

If only a list is passed to cov.x, a matrix of Distances is returned
}
\description{
Calculates distances between covariance matrices.
}
\examples{
c1 <- RandomMatrix(10)
c2 <- RandomMatrix(10)
c3 <- RandomMatrix(10)
MatrixDistance(c1, c2)

MatrixDistance(list(c1, c2, c3))


c4 <- RandomMatrix(10)
MatrixDistance(list(c1, c2, c3), c4)
}
\author{
Diogo Melo
}
\seealso{
\code{\link{RiemannDist}},\code{\link{OverlapDist}}
}
\keyword{matrixcomparison}
\keyword{matrixdistance}

