% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MatrixOverlapDist}
\alias{MatrixOverlapDist}
\title{Distribution overlap distance}
\usage{
MatrixOverlapDist(cov.x, cov.y, iterations = 10000)
}
\arguments{
\item{cov.x}{covariance or correlation matrix}

\item{cov.y}{covariance or correlation matrix}

\item{iterations}{number of drows}
}
\value{
Overlap distance between cov.x and cov.y
}
\description{
Calculates the overlap between two normal distributions,
defined as the probability that a draw from one distribution
comes from the other
}

