% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixStatistics.R
\name{MeanMatrixStatistics}
\alias{MeanMatrixStatistics}
\alias{Autonomy}
\alias{ConditionalEvolvability}
\alias{Constraints}
\alias{Evolvability}
\alias{Flexibility}
\alias{Pc1Percent}
\alias{Respondability}
\title{Calculate mean values for various matrix statistics}
\usage{
MeanMatrixStatistics(
  cov.matrix,
  iterations = 1000,
  full.results = FALSE,
  parallel = FALSE,
  beta.mat = NULL,
  include.R2 = FALSE,
  include.ICV = FALSE
)

Autonomy(cov.matrix, beta.mat = NULL, iterations = 1000)

ConditionalEvolvability(cov.matrix, beta.mat = NULL, iterations = 1000)

Constraints(cov.matrix, beta.mat = NULL, iterations = 1000)

Evolvability(cov.matrix, beta.mat = NULL, iterations = 1000)

Flexibility(cov.matrix, beta.mat = NULL, iterations = 1000)

Pc1Percent(cov.matrix)

Respondability(cov.matrix, beta.mat = NULL, iterations = 1000)
}
\arguments{
\item{cov.matrix}{A covariance matrix}

\item{iterations}{Number of random vectors to be used in calculating the stochastic statistics when beta.mat is NULL.}

\item{full.results}{If TRUE, full distribution of statistics will be returned.}

\item{parallel}{if TRUE computations are done in parallel. Some foreach backend must be registered, like doParallel or doMC.}

\item{beta.mat}{Matrix of direction vectors used to calculate the stochastic statistics. If NULL, a random matrix of normalized vectors is generated.}

\item{include.R2}{Logical. If TRUE, include the legacy mean squared correlation metric in the returned output.}

\item{include.ICV}{Logical. If TRUE, include the legacy ICV metric in the returned output.}
}
\value{
dist Full distribution of stochastic statistics, only if full.resuts == TRUE

mean Mean value for all statistics
}
\description{
Calculates: Mean Squared Correlation, ICV, Autonomy, ConditionalEvolvability, Constraints, Evolvability, Flexibility, Pc1Percent, Respondability.
}
\examples{
cov.matrix <- cov(iris[,1:4])
MeanMatrixStatistics(cov.matrix)

\dontrun{
#Multiple threads can be used with some foreach backend library, like doMC or doParallel
library(doMC)
registerDoMC(cores = 2)
MeanMatrixStatistics(cov.matrix, parallel = TRUE)
}
}
\references{
Hansen, T. F., and Houle, D. (2008). Measuring and comparing evolvability
and constraint in multivariate characters. Journal of evolutionary
biology, 21(5), 1201-19. doi:10.1111/j.1420-9101.2008.01573.x
}
\author{
Diogo Melo Guilherme Garcia
}
\keyword{Autonomy}
\keyword{ConditionalEvolvability}
\keyword{Constraints}
\keyword{Evolvability}
\keyword{Flexibility}
\keyword{Pc1Percent}
\keyword{Respondability}
