% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModularityGamma.R
\name{ModularityGamma}
\alias{ModularityGamma}
\title{Calculates Gamma-distance between modular hypothesis and covariance matrix}
\usage{
ModularityGamma(cor.matrix, modularity.hypot)
}
\arguments{
\item{cor.matrix}{Correlation matrix}

\item{modularity.hypot}{Matrix of hypothesis. Each line represents a trait and each column a module.
if modularity.hypot[i,j] == 1, trait i is in module j.}
}
\description{
Combines and compares many modularity hypothesis to a covariance matrix. Comparison values are
ajusted to the number os zeros in the hypothesis using a linear regression.
}
\note{
Hypothesis can be named as column names, and these will be used to make labels in the 
output
}
\examples{
# Create a single modularity hypothesis:
hypot = rep(c(1, 0), each = 6)
cor.hypot = CreateHypotMatrix(hypot)

# First with an unstructured matrix:
un.cor = RandomMatrix(12)
MantelModTest(cor.hypot, un.cor)

# Now with a modular matrix:
modules = matrix(c(rep(c(1, 0, 0), each = 5),
rep(c(0, 1, 0), each = 5),
rep(c(0, 0, 1), each = 5)), 15)
cor.hypot = CreateHypotMatrix(modules)[[4]]
hypot.mask = matrix(as.logical(cor.hypot), 15, 15)
mod.cor = matrix(NA, 15, 15)
mod.cor[ hypot.mask] = runif(length(mod.cor[ hypot.mask]), 0.8, 0.9) # within-modules
mod.cor[!hypot.mask] = runif(length(mod.cor[!hypot.mask]), 0.3, 0.4) # between-modules
diag(mod.cor) = 1
mod.cor = (mod.cor + t(mod.cor))/2 # correlation matrices should be symmetric

# True hypothesis and a bunch of random ones.
hypothetical.modules = cbind(modules, matrix(sample(c(1, 0), 4*15, replace=TRUE), 15, 4))

# if hypothesis columns are not named they are assigned numbers
colnames(hypothetical.modules) <- letters[1:7]
 
ModularityGamma(mod.cor, hypothetical.modules)
}
\references{
Marquez, E.J. 2008. A statistical framework for testing modularity in multidimensional data. Evolution 62:2688-2708.
}

