\name{Morphometrics-package}
\alias{Morphometrics-package}
\alias{Morphometrics}
\docType{package}
\title{
Morphometrics functions for matrix estimation and comparisons
}
\description{
Resonable implementations for a few morphometrics techniques.
}
\details{
\tabular{ll}{
Package: \tab Morphometrics\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-09-16\cr
License: \tab MIT\cr
}
Random Skewers is pretty good
}
\author{
Maintainer: Diogo Melo <diogro@cecm.usp.br>
Diogo Melo, Guilherme Garcia
}
\references{
    Marroig, Melo, Garcia 2012
    Marroig et al 2011
}
\keyword{Morphometrics}
\examples{
    RandomMatrix(10)
}
