% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{Normalize}
\alias{Normalize}
\alias{Norm}
\title{Normalize and Norm}
\usage{
Normalize(x)

Norm(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Normalized vector or inpout vector norm.
}
\description{
Norm returns the euclidian norm of a vector, Normalize returns a vector with unit norm.
}
\examples{
x <- rnorm(10)
n.x <- Normalize(x)
Norm(x)
Norm(n.x)
}
\author{
Diogo Melo, Guilherme Garcia
}
