% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDistance.R
\name{OverlapDist}
\alias{OverlapDist}
\title{Distribution overlap distance}
\usage{
OverlapDist(cov.x, cov.y, iterations = 10000)
}
\arguments{
\item{cov.x}{covariance or correlation matrix}

\item{cov.y}{covariance or correlation matrix}

\item{iterations}{number of drows}
}
\value{
Overlap distance between cov.x and cov.y
}
\description{
Calculates the overlap between two normal distributions, 
defined as the probability that a draw from one distribution 
comes from the other
}
\references{
Ovaskainen, O. (2008). A Bayesian framework for comparative quantitative genetics. Proceedings of the Royal Society B, 669-678. doi:10.1098/rspb.2007.0949
}
