% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHypotMatrix.R
\name{Partition2HypotMatrix}
\alias{Partition2HypotMatrix}
\title{Create binary hypothesis}
\usage{
Partition2HypotMatrix(x)
}
\arguments{
\item{x}{vector of trait partition. Each partition receive the same symbol.}
}
\value{
Matrix of hypothesis. Each line represents a trait and each column a module.
if modularity.hypot[i,j] == 1, trait i is in module j.
}
\description{
Takes a vetor describing a trait partition and returns a binary matrix of the partitions where each line represents a trait and each column a module. In the output matrix, if modularity.hypot[i,j] == 1, trait i is in module j.
}
\examples{
x = sample(c(1, 2, 3), 10, replace = TRUE)
Partition2HypotMatrix(x) 
}
