% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PhyloCompare}
\alias{PhyloCompare}
\title{Compares sister groups}
\usage{
PhyloCompare(tree, node.data, CompareFunc = RandomSkewers)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{node.data}{list of node data}

\item{CompareFunc}{comparison function}
}
\value{
list with calculated comparisons for each node, using labels or numbers from tree
}
\description{
Calculates the comparison of some statistic between sister groups along a phylogeny
}
\note{
Phylogeny must be fully resolved
}
\examples{
library(ape)
data(bird.orders)
tree <- bird.orders
mat.list <- RandomMatrix(5, length(tree$tip.label))
names(mat.list) <- tree$tip.label
sample.sizes <- runif(length(tree$tip.label), 15, 20)
phyl.state <- AncestralStates(tree, mat.list, sample.sizes)
PhyloCompare(tree, phyl.state)
}

