% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RandomCorrelation.R
\name{RandomCorrelation}
\alias{RandomCorrelation}
\title{Random Correlation}
\usage{
RandomCorrelation(dimension, k = 10^-3)
}
\arguments{
\item{num.traits}{Number of traits in random matrix}

\item{ke}{Parameter for correlation matrix generation. Involves check for positive defitness}
}
\value{
Random correlation matrix with size (dimension x dimension)
}
\description{
Create a random correlation matrix using RcppEigen
}
\author{
Edgar Zanella
}
\keyword{randommatrices}

