% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvolQG.R
\docType{data}
\name{ratones}
\alias{ratones}
\title{Linear distances for five mouse lines}
\format{
data.frame
}
\source{
Dryad Archive \doi{10.5061/dryad.5gr8r}
}
\usage{
data(ratones)
}
\description{
Skull distances measured from landmarks in 5 mice lines: 4 body weight selection lines and 1 control line. 
Originally published in Penna, A., Melo, D. et. al (2017) \doi{10.1111/evo.13304}
}
\examples{
data(ratones)
   
# Estimating a W matrix, controlling for line and sex
model_formula = paste0("cbind(", 
                       paste(names(ratones)[13:47], collapse = ", "),
                       ") ~ SEX + LIN")
ratones_W_model = lm(model_formula, data = ratones)
W_matrix = CalculateMatrix(ratones_W_model)

# Estimating the divergence between the two direction of selection
delta_Z = colMeans(ratones[ratones$selection == "upwards", 13:47]) -
          colMeans(ratones[ratones$selection == "downwards", 13:47])
          
 # Reconstructing selection gradients with and without noise control         
Beta = solve(W_matrix, delta_Z)
Beta_non_noise = solve(ExtendMatrix(W_matrix, ret.dim = 10)$ExtMat, delta_Z)

# Comparing the selection gradients to the observed divergence
Beta \%*\% delta_Z /(Norm(Beta) * Norm(delta_Z))
Beta_non_noise \%*\% delta_Z /(Norm(Beta_non_noise) * Norm(delta_Z))      
          
}
\references{
Penna, A., Melo, D., Bernardi, S., Oyarzabal, M.I. and Marroig, G. (2017), The evolution of phenotypic integration: How directional selection reshapes covariation in mice. Evolution, 71: 2370-2380. \doi{10.1111/evo.13304}
}
\keyword{datasets}
