% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeEigenanalysis.R
\name{RelativeEigenanalysis}
\alias{RelativeEigenanalysis}
\title{Relative Eigenanalysis}
\usage{
RelativeEigenanalysis(cov.x, cov.y, symmetric = FALSE)
}
\arguments{
\item{cov.x}{covariance matrix}

\item{cov.y}{covariance matrix}

\item{symmetric}{compute symmetric eigenanalysis?}
}
\value{
list with two objects: eigenvalues and eigenvectors
}
\description{
Computes relative eigenvalues and eigenvectors between a pair of covariance matrices.
}
\references{
Bookstein, Fred L., and Philipp Mitteroecker. "Comparing Covariance Matrices by
Relative Eigenanalysis, with Applications to Organismal Biology." Evolutionary Biology 41, no. 2
(June 1, 2014): 336–50. doi:10.1007/s11692-013-9260-5.
}
\author{
Guilherme Garcia, Diogo Melo
}
