% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevertMatrix.R
\name{RevertMatrix}
\alias{RevertMatrix}
\title{Revert Matrix}
\usage{
RevertMatrix(values, etd, scaled = TRUE)
}
\arguments{
\item{values}{vector of values to build matrix, each value corresponding 
to a score on the ordered set of eigentensors (up to the maximum number of 
eigentensors on the target decomposition)}

\item{etd}{Eigentensor decomposition of m covariance matrices for 
k traits (obtained from \code{\link{EigenTensorDecomposition}})}

\item{scaled}{should we treat each score as a value given in standard 
deviations for each eigentensor? Defaults to TRUE}
}
\value{
A symmetric covariance matrix with k traits
}
\description{
Constructs a covariance matrix based on scores over covariance matrix eigentensors.
}
\examples{

}
\references{
Basser P. J., Pajevic S. 2007. Spectral decomposition of a 4th-order 
covariance tensor: Applications to diffusion tensor MRI. Signal Processing. 87:220–236.

Hine E., Chenoweth S. F., Rundle H. D., Blows M. W. 2009. Characterizing 
the evolution of genetic variance using genetic covariance tensors. Philosophical 
transactions of the Royal Society of London. Series B, Biological sciences. 364:1567–78.
}

