% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RiemannDist}
\alias{RiemannDist}
\title{Matrix Riemann Distance}
\usage{
RiemannDist(cov.x, cov.y)
}
\arguments{
\item{cov.x}{covariance or correlation matrix}

\item{cov.y}{covariance or correlation matrix}
}
\value{
Riemann distance between cov.x and cov.y
}
\description{
Return distance between two covariance matrices
}
\author{
Edgar Zanella
}
\keyword{matrixcomparison}
\keyword{matrixdistance}

