% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localShapeVariables.R
\name{Rotate2MidlineMatrix}
\alias{Rotate2MidlineMatrix}
\title{Midline rotate}
\usage{
Rotate2MidlineMatrix(X, midline)
}
\arguments{
\item{X}{shape array}

\item{midline}{rows for the midline landmarks}
}
\value{
Rotation matrix
}
\description{
Returns the rotation matrix that aligns a specimen sagital line
to plane y = 0 (2D) or z = 0 (3D)
}
\author{
Guilherme Garcia
}
