% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localShapeVariables.R
\name{TPS}
\alias{TPS}
\title{TPS transform}
\usage{
TPS(target.shape, reference.shape)
}
\arguments{
\item{target.shape}{Target shape}

\item{reference.shape}{Reference shape}
}
\value{
A list with the transformation parameters and a function that gives
the value of the TPS function at each point for numerical differentiation
}
\description{
Calculates the Thin Plate Spline transform between a reference shape and a target shape
}
\author{
Guilherme Garcia
}
