% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BootstrapRep.R
\name{BootstrapRep}
\alias{BootstrapRep}
\title{Bootstrap analysis via resampling}
\usage{
BootstrapRep(ind.data, ComparisonFunc, ..., iterations = 1000,
  correlation = FALSE, num.cores = 1)
}
\arguments{
\item{ind.data}{Matrix of residuals or indiviual measurments}

\item{ComparisonFunc}{comparison function}

\item{...}{Aditional arguments passed to ComparisonFunc}

\item{iterations}{Number of resamples to take}

\item{correlation}{If TRUE, correlation matrix is used, else covariance matrix.}

\item{num.cores}{Number of threads to use in computation.
The doMC library must be loaded.}
}
\value{
returns the mean repeatability, that is, the mean value of comparisons from samples to original statistic.
}
\description{
Calculates the repeatability of the covariance matrix of the suplied data
  via bootstrap resampling
}
\details{
Samples with replacement are taken from the full population, a statistic calculated
  and compared to the full population statistic.
}
\examples{
BootstrapRep(iris[,1:4], MantelCor, iterations = 5, correlation = TRUE)

BootstrapRep(iris[,1:4], RandomSkewers, iterations = 50)

BootstrapRep(iris[,1:4], KrzCor, iterations = 50, correlation = TRUE)

BootstrapRep(iris[,1:4], PCAsimilarity, iterations = 50)

#Multiple threads can be used with doMC library
library(doMC)
BootstrapRep(iris[,1:4], PCAsimilarity,
             iterations = 5,
             num.cores = 2)
}
\author{
Diogo Melo, Guilherme Garcia
}
\seealso{
\code{\link{MonteCarloStat}}, \code{\link{AlphaRep}}
}
\keyword{bootstrap}
\keyword{repetabilities}

