% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DriftTest.R
\name{DriftTest}
\alias{DriftTest}
\title{Test drift hypothesis}
\usage{
DriftTest(means, cov.matrix, show.plot = TRUE)
}
\arguments{
\item{means}{list or array of species means being compared. array must have means in the rows.}

\item{cov.matrix}{ancestral covariance matrix for all populations}

\item{show.plot}{boolean. If TRUE, plot of eigenvalues of ancetral matrix by between group variance is showed.}
}
\value{
list os results
}
\description{
Given a set of covariance matrices and means for terminals, test the hypothesis
that obseved divergency is larger/smaller than expected by drift alone.
}
\examples{
means = array(rnorm(40*10), c(10, 40))
cov.matrix = RandomMatrix(40, 1, 1, 10)
DriftTest(means, cov.matrix)
}
\author{
Ana Paula Assis, Diogo Melo
}
\references{
Marroig, G., & Cheverud, J. M. (2004). Did natural selection or genetic drift
produce the cranial diversification of neotropical monkeys? The American Naturalist, 163(3), 417-428. doi:10.1086/381693
}

