% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrzSubspaceBootstrap.R
\name{KrzSubspaceBootstrap}
\alias{KrzSubspaceBootstrap}
\title{Quasi-Bayesian Krzanowski subspace comparison}
\usage{
KrzSubspaceBootstrap(x, rep = 1, MCMCsamples = 1000, parallel = FALSE)
}
\arguments{
\item{x}{list of linear models from which P-matrices should be calculated}

\item{rep}{number of bootstrap samples to be made}

\item{MCMCsamples}{number of MCMCsamples for each P-matrix posterior distribution.}

\item{parallel}{if TRUE computations are done in parallel. Some foreach backend must be registered, like doParallel or doMC.}
}
\description{
Calculates the usual Krzanowski subspace comparison using a posterior samples
for a set of phenotypic covariance matrices. Then, this observed comparison 
is contrasted to the subspace comparison across a permutation of the original
data. Residuals, which are used to calculate the observed P-matrices, are 
shuffled across groups. This process is repeated, creating a null distribution
of  subspace comparisons under the hypothesis that all P-matrices come from the
same population. This method is a modification on the fully Bayesian method 
proposed in Aguirre et. al 2014 and improved in Morrisey et al 2020.
}
