% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/l_modularity.R
\name{LModularity}
\alias{LModularity}
\title{L Modularity}
\usage{
LModularity(cor.matrix)
}
\arguments{
\item{cor.matrix}{correlation matrix}
}
\value{
List with L-Modularity value and trait partition
}
\description{
Calculates the L-Modularity (Newman-type modularity) and the partition of traits that minimizes L-Modularity
}
\examples{
cor.matrix = RandomMatrix(10)
LModularity(cor.matrix)
}
\references{
Modularity and community structure in networks (2006) M. E. J. Newman,  8577-8582, doi: 10.1073/pnas.0601602103
}

