% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixCompare.R
\name{MatrixCompare}
\alias{MatrixCompare}
\title{Matrix Compare}
\usage{
MatrixCompare(cov.x, cov.y, id = ".id")
}
\arguments{
\item{cov.x}{covariance or correlation matrix}

\item{cov.y}{covariance or correlation matrix}

\item{id}{name of the comparison column}
}
\value{
data.frame of comparisons
}
\description{
Compare two matrices using all available methods. Currently RandomSkewers, MantelCor, KrzCor and PCASimilarity
}
\examples{

cov.x = RandomMatrix(10, 1, 1, 10)
cov.y = RandomMatrix(10, 1, 10, 20)
MatrixCompare(cov.x, cov.y)
}
