\name{MatrixRiemannDist}
\alias{MatrixRiemannDist}
\title{Matrix Riemann Distance}
\usage{
MatrixRiemannDist(cov.x, cov.y)
}
\arguments{
  \item{cov.x}{covariance or correlation matrix}

  \item{cov.y}{covariance or correlation matrix}
}
\description{
Return distance between two covariance matrices
}
\keyword{matrixcomparison}
\keyword{matrixdistance}

