% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MonteCarloStat.R
\name{MonteCarloRep}
\alias{MonteCarloRep}
\title{Parametric repeatabilities with covariance or correlation matrices}
\usage{
MonteCarloRep(cov.matrix, ComparisonFunc, sample.size, ..., iterations = 1000,
  correlation = FALSE, num.cores = 1)
}
\arguments{
\item{cov.matrix}{Covariance matrix.}

\item{ComparisonFunc}{comparison function}

\item{sample.size}{Size of the random populations}

\item{...}{Aditional arguments passed to CompareFunc}

\item{iterations}{Number of random populations}

\item{correlation}{If TRUE, correlation matrix is used, else covariance matrix. MantelCor always uses correlation matrix.}

\item{num.cores}{If list is passed, number of threads to use in computation.
The doMC library must be loaded.}
}
\value{
returns the mean repeatability, or mean value of comparisons from samples to original statistic.
}
\description{
Using a multivariate normal model, random populations are generated
using the suplied covariance matrix. A statistic is calculated on the
random population and compared to the statistic calculated on the
original matrix.
}
\details{
Since this function uses multivariate normal model to generate populations, only covariance matrices should be used.
}
\examples{
cov.matrix <- RandomMatrix(10, 1, 1, 10)

MonteCarloRep(cov.matrix, RandomSkewers, 30, iterations = 50)
MonteCarloRep(cov.matrix, RandomSkewers, 30, num.vectors = 100, iterations = 50, correlation = TRUE)
MonteCarloRep(cov.matrix, MantelCor, 30, permutations = 1, correlation = TRUE)
MonteCarloRep(cov.matrix, KrzCor, 30)
MonteCarloRep(cov.matrix, KrzCor, 30, correlation = TRUE)

#Multiple threads can be used with doMC library
library(doMC)
MonteCarloRep(cov.matrix, RandomSkewers, 30, iterations = 100, num.cores = 2)

#Creating repeatability vector for a list of matrices
mat.list <- RandomMatrix(10, 3, 1, 10)
laply(mat.list, MonteCarloRep, KrzCor, 30, correlation = TRUE)
}
\author{
Diogo Melo Guilherme Garcia
}
\seealso{
\code{\link{BootstrapRep}}, \code{\link{AlphaRep}}
}
\keyword{montecarlo}
\keyword{parametricsampling}
\keyword{repeatability}

