% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrzSubspaceBootstrap.R
\name{PlotKrzSubspace}
\alias{PlotKrzSubspace}
\title{Plot KrzSubspace boostrap comparison}
\usage{
PlotKrzSubspace(x)
}
\arguments{
\item{x}{output from KrzSubspaceDataFrame() function.}
}
\value{
ggplot2 object with the observed vs. random eigenvalues mean and posterior confidence intervals
}
\description{
Shows the null and observed distribution of eigenvalues from the Krzanowski subspace comparison
}
