% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PrintMatrix.R
\name{PrintMatrix}
\alias{PrintMatrix}
\alias{PrintMatrix.default}
\alias{PrintMatrix.list}
\title{Print Matrix to file}
\usage{
PrintMatrix(x, ...)

\method{PrintMatrix}{default}(x, out.file = "./matrix.csv", ...)

\method{PrintMatrix}{list}(x, out.file = "./matrix.csv", ...)
}
\arguments{
\item{x}{Matrix or list of matrices}

\item{...}{Aditional parameters}

\item{out.file}{Output file}
}
\value{
Prints coma separated matrices, with labels
}
\description{
Print a matrix or a list of matrices to file
}
\examples{
m.list <- RandomMatrix(10, 4)
PrintMatrix(m.list)
}
\author{
Diogo Melo
}

