% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintMatrix.R
\name{PrintMatrix}
\alias{PrintMatrix}
\alias{PrintMatrix.default}
\alias{PrintMatrix.list}
\title{Print Matrix to file}
\usage{
PrintMatrix(x, ...)

\method{PrintMatrix}{default}(x, output.file, ...)

\method{PrintMatrix}{list}(x, output.file, ...)
}
\arguments{
\item{x}{Matrix or list of matrices}

\item{...}{Additional parameters}

\item{output.file}{Output file}
}
\value{
Prints coma separated matrices, with labels
}
\description{
Print a matrix or a list of matrices to file
}
\examples{
m.list <- RandomMatrix(10, 4)
tmp = file.path(tempdir(), "matrix.csv")
PrintMatrix(m.list, output.file = tmp )
}
\author{
Diogo Melo
}
