% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintMatrix.R
\name{PrintMatrix}
\alias{PrintMatrix}
\alias{PrintMatrix.default}
\alias{PrintMatrix.list}
\title{Print Matrix to file}
\usage{
PrintMatrix(x, ...)

\method{PrintMatrix}{default}(x, output.file = "./matrix.csv", ...)

\method{PrintMatrix}{list}(x, output.file = "./matrix.csv", ...)
}
\arguments{
\item{x}{Matrix or list of matrices}

\item{...}{Aditional parameters}

\item{output.file}{Output file}
}
\value{
Prints coma separated matrices, with labels
}
\description{
Print a matrix or a list of matrices to file
}
\examples{
\dontrun{
m.list <- RandomMatrix(10, 4)
PrintMatrix(m.list)}
}
\author{
Diogo Melo
}
