% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomCorr.R
\name{RandCorr}
\alias{RandCorr}
\title{Random correlation matrix}
\usage{
RandCorr(num.traits, ke = 10^-3)
}
\arguments{
\item{num.traits}{Number of traits in random matrix}

\item{ke}{Parameter for correlation matrix generation. Involves check for positive defitness}
}
\value{
Random Matrix
}
\description{
Internal function for generating random correlation matrices.
Use RandomMatrix() instead.
}
\author{
Diogo Melo Edgar Zanella
}
\keyword{randommatrices}
