% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RandomCorrelation.R
\name{RandomCorrelation}
\alias{RandomCorrelation}
\title{Random correlation matrix}
\usage{
RandomCorrelation(dimension, k = 10^-3)
}
\arguments{
\item{dimension}{Number of traits in random matrix}

\item{k}{Parameter for correlation matrix generation. Involves check for positive defitness}
}
\value{
Random correlation matrix with size (dimension x dimension)
}
\description{
Internal function for generating random correlation matrices.
Use RandomMatrix() instead.
}
\author{
Edgar Zanella
}
\keyword{randommatrices}

