% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDistance.R
\name{RiemannDist}
\alias{RiemannDist}
\title{Matrix Riemann distance}
\usage{
RiemannDist(cov.x, cov.y)
}
\arguments{
\item{cov.x}{covariance or correlation matrix}

\item{cov.y}{covariance or correlation matrix}
}
\value{
Riemann distance between cov.x and cov.y
}
\description{
Return distance between two covariance matrices
}
\references{
Mitteroecker, P., & Bookstein, F. (2009). The ontogenetic trajectory of the phenotypic covariance matrix, with examples from craniofacial shape in rats and humans. Evolution, 63(3), 727-737. doi:10.1111/j.1558-5646.2008.00587.x
}
\author{
Edgar Zanella
}
\keyword{matrixDistance}
\keyword{matrixcomparison}
